+++ lib/ruby_smart/support/gem_version.rb
@@ -9,8 +9,8 @@ module RubySmart
 
     module VERSION
       MAJOR = 1
-      MINOR = 1
-      TINY  = 1
+      MINOR = 2
+      TINY  = 0
       PRE   = nil
 
       STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")